requirements.txt
streamlit==1.29.0
numpy==1.23.5
scikit-learn==1.2.2
tensorflow-cpu==2.11.0
librosa==0.10.0.post2
soundfile==0.12.1
pydub==0.25.1
